# src/docs/swagger.yaml
basePath: /api/v1
definitions:
...
  model.CreateProjectRequest:
    properties:
      description:
        type: string
      name:
        type: string
      workspaceId:
        description: optional workspace to assign project to
        type: string
    required:
    - name
    type: object
...
  model.Project:
    properties:
      created_at:
        type: string
      description:
        type: string
      id:
        type: integer
      name:
        type: string
      nsid:
        description: Namespace ID
        type: string
      updated_at:
        type: string
      workspaces:
        description: M:N relationship
        items:
          $ref: '#/definitions/model.Workspace'
        type: array
    type: object
  model.ProjectSyncApplyAssignedItem:
    properties:
      id:
        type: integer
      name:
        type: string
      nsId:
        type: string
    type: object
  model.ProjectSyncApplyCreatedItem:
    properties:
      id:
        type: integer
      name:
        type: string
      nsId:
        type: string
    type: object
  model.ProjectSyncApplyFailedItem:
    properties:
      error:
        type: string
      nsId:
        type: string
    type: object
  model.ProjectSyncApplyRequest:
    properties:
      nsIds:
        items:
          type: string
        type: array
      workspaceId:
        type: string
    required:
    - nsIds
    - workspaceId
    type: object
  model.ProjectSyncApplyResponse:
    properties:
      assigned:
        items:
          $ref: '#/definitions/model.ProjectSyncApplyAssignedItem'
        type: array
      created:
        items:
          $ref: '#/definitions/model.ProjectSyncApplyCreatedItem'
        type: array
      failed:
        items:
          $ref: '#/definitions/model.ProjectSyncApplyFailedItem'
        type: array
      skipped:
        items:
          $ref: '#/definitions/model.ProjectSyncApplySkippedItem'
        type: array
    type: object
  model.ProjectSyncApplySkippedItem:
    properties:
      nsId:
        type: string
      reason:
        type: string
    type: object
  model.ProjectSyncDiffMissingItem:
    properties:
      description:
        type: string
      name:
        type: string
      nsId:
        type: string
    type: object
  model.ProjectSyncDiffResponse:
    properties:
      missingProjects:
        items:
          $ref: '#/definitions/model.ProjectSyncDiffMissingItem'
        type: array
      unassignedProjects:
        items:
          $ref: '#/definitions/model.ProjectSyncDiffUnassignedItem'
        type: array
    type: object
  model.ProjectSyncDiffUnassignedItem:
    properties:
      description:
        type: string
      id:
        type: integer
      name:
        type: string
      nsId:
        type: string
    type: object
...
  model.Workspace:
    properties:
      created_at:
        type: string
      description:
        type: string
      id:
        type: integer
      name:
        type: string
      projects:
        items:
          $ref: '#/definitions/model.Project'
        type: array
      updated_at:
        type: string
    type: object
...
  model.WorkspaceProjectMappingRequest:
    properties:
      projectIds:
        items:
          type: string
        type: array
      workspaceId:
        type: string
    required:
    - projectIds
    - workspaceId
    type: object
...
host: localhost
info:
  contact: {}
  description: MC IAM Manager API Documentation
  title: MC IAM Manager API
  version: 0.6.2
paths:
...
  /api/projects:
    post:
      consumes:
      - application/json
      description: Create a new project with the specified information. Optionally
        specify a workspace to assign the project to.
      operationId: createProject
      parameters:
      - description: Project Info
        in: body
        name: project
        required: true
        schema:
          $ref: '#/definitions/model.CreateProjectRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/model.Project'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Create new project
      tags:
      - projects
  /api/projects/assign/workspaces:
    post:
      consumes:
      - application/json
      description: 프로젝트에 워크스페이스를 연결합니다.
      operationId: addWorkspaceToProject
      parameters:
      - description: Workspace and Project IDs
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/model.WorkspaceProjectMappingRequest'
      produces:
      - application/json
      responses:
        "204":
          description: No Content
        "400":
          description: 'error: 잘못된 ID 형식'
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: 'error: 프로젝트 또는 워크스페이스를 찾을 수 없습니다'
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: 'error: 서버 내부 오류'
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: 프로젝트에 워크스페이스 연결
      tags:
      - projects
  /api/projects/id/{projectId}:
    delete:
      consumes:
      - application/json
      description: Delete a project by its ID.
      operationId: deleteProject
      parameters:
      - description: Project ID
        in: path
        name: projectId
        required: true
        type: string
      produces:
      - application/json
      responses:
        "204":
          description: No Content
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Delete project
      tags:
      - projects
    get:
      consumes:
      - application/json
      description: Retrieve project details by project ID.
      operationId: getProjectByID
      parameters:
      - description: Project ID
        in: path
        name: projectId
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.Project'
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Get project by ID
      tags:
      - projects
    put:
      consumes:
      - application/json
      description: Update the details of an existing project.
      operationId: updateProject
      parameters:
      - description: Project ID
        in: path
        name: projectId
        required: true
        type: string
      - description: Project Info
        in: body
        name: project
        required: true
        schema:
          $ref: '#/definitions/model.Project'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.Project'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Update project
      tags:
      - projects
  /api/projects/id/{projectId}/workspaces:
    get:
      consumes:
      - application/json
      description: Retrieve list of workspaces that the project is assigned to
      operationId: getProjectWorkspaces
      parameters:
      - description: Project ID
        in: path
        name: projectId
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/model.Workspace'
            type: array
        "400":
          description: 'error: Invalid project ID'
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: 'error: Project not found'
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: 'error: Internal server error'
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Get workspaces assigned to project
      tags:
      - projects
  /api/projects/list:
    post:
      consumes:
      - application/json
      description: Retrieve a list of all projects.
      operationId: listProjects
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/model.Project'
            type: array
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: List all projects
      tags:
      - projects
  /api/projects/name/{projectName}:
    get:
      consumes:
      - application/json
      description: Get project details by name
      operationId: getProjectByName
      parameters:
      - description: Project Name
        in: path
        name: name
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.Project'
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Get project by name
      tags:
      - projects
  /api/projects/unassign/workspaces:
    delete:
      consumes:
      - application/json
      description: Remove a workspace from a project
      operationId: removeWorkspaceFromProject
      parameters:
      - description: Workspace and Project IDs
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/model.WorkspaceProjectMappingRequest'
      produces:
      - application/json
      responses:
        "204":
          description: No Content
        "400":
          description: 'error: Invalid request'
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: 'error: Internal server error'
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Remove workspace from project
      tags:
      - projects
...
  /api/setup/projects/sync:
    post:
      consumes:
      - application/json
      description: 지정된 nsId 목록의 project를 생성하거나 지정된 workspace에 할당합니다.
      operationId: applyProjectSync
      parameters:
      - description: Sync Apply Request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/model.ProjectSyncApplyRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.ProjectSyncApplyResponse'
        "400":
          description: 'error: 잘못된 요청'
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: 'error: 서버 내부 오류'
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: 선택한 namespace를 특정 workspace에 동기화
      tags:
      - projects
  /api/setup/projects/sync-diff:
    get:
      description: mc-infra-manager의 namespace 목록과 로컬 프로젝트를 비교하여 불일치 목록을 반환합니다. DB
        변경 없음.
      operationId: getProjectSyncDiff
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.ProjectSyncDiffResponse'
        "500":
          description: 'error: 서버 내부 오류'
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: mc-infra-manager와 프로젝트 동기화 차이 조회
      tags:
      - projects
...
  /api/setup/sync-projects:
    post:
      consumes:
      - application/json
      description: mc-infra-manager의 네임스페이스 목록을 조회하여 로컬 DB에 없는 프로젝트를 추가합니다.
      operationId: syncProjects
      produces:
      - application/json
      responses:
        "200":
          description: 'message: Project synchronization successful'
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: 'error: 서버 내부 오류 또는 동기화 실패'
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: mc-infra-manager와 프로젝트 동기화
      tags:
      - projects
...
  /api/workspaces/assign/projects:
    post:
      consumes:
      - application/json
      description: Add a project to a workspace
      operationId: addProjectToWorkspace
      parameters:
      - description: Workspace ID
        in: path
        name: id
        required: true
        type: string
      - description: Project ID
        in: path
        name: projectId
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.Workspace'
        "400":
          description: 'error: Invalid request'
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: 'error: Unauthorized'
          schema:
            additionalProperties:
              type: string
            type: object
        "403":
          description: 'error: Forbidden'
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: 'error: Workspace or Project not found'
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Add project to workspace
      tags:
      - workspaces
...
  /api/workspaces/unassign/projects:
    delete:
      consumes:
      - application/json
      description: Remove a project from a workspace
      operationId: removeProjectFromWorkspace
      parameters:
      - description: Workspace ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses: {}
      summary: Remove project from workspace
      tags:
      - workspaces
...
securityDefinitions:
  BearerAuth:
    description: Type "Bearer" followed by a space and JWT token.
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"